requirements.txt
astroid==3.0.3
bcrypt==4.1.2
black==23.12.1
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.7
coverage==7.11.3
cryptography==41.0.7
dill==0.4.0
dnspython==2.8.0
email-validator==2.1.0
exceptiongroup==1.3.0
flake8==6.1.0
Flask==3.0.0
Flask-Cors==4.0.0
greenlet==3.2.4
gunicorn==21.2.0
idna==3.11
iniconfig==2.3.0
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pluggy==1.6.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.23
pyflakes==3.1.0
PyJWT==2.8.0
pylint==3.0.3
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
requests==2.31.0
six==1.17.0
SQLAlchemy==2.0.23
tomli==2.3.0
tomlkit==0.13.3
typing_extensions==4.15.0
urllib3==2.1.0
validators==0.22.0
waitress==2.1.2
Werkzeug==3.0.1
